// src/components/__generated__/enhancedPostFeedQuery.graphql.ts
export type enhancedPostFeedQuery$variables = {
  first: number;
};
export type enhancedPostFeedQuery$data = {
  readonly timeline: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"postCard">;
      };
    }>;
  };
};
export type enhancedPostFeedQuery = {
  response: enhancedPostFeedQuery$data;
  variables: enhancedPostFeedQuery$variables;
};